Disable swipe between tabs
Through a CustomViexPager, this adds the possibility to disable swiping between tabs.
I think it's better to disable it. So, we'll be able to add so swipe action in the tasks' tab like swipe to left to move backward and swipe to right to move forward. And have more space for the task itself.